+++ lib/resources/app_vectors.dart
@@ -4,4 +4,7 @@ class AppVectors {
   AppVectors._();
 
   static const String cityIcon = 'assets/vectors/city_icon.svg';
+  static const String comment = 'assets/vectors/comment.svg';
+  static const String heartFilled = 'assets/vectors/heart_filled.svg';
+  static const String heartOutlined = 'assets/vectors/heart_outlined.svg';
 }
